Grabbing the cartridges that are only visible inside cone
It works. I set up a third table to the left. The cartridges on the left side of the table have no invertedMask material, the ones on the right side of the table do have. The cartridges in the middle of each row are visible in the cone always. The ones in front and behind that cartridge only work with a specific cartridge that has to be inserted into the lantern.